Update README.md
Added Maps as a fluid resource. This is used by the Tinkerer playbook.

Fixed inconsistent variable name for the Rain's Rain fluid resource. Updating the sheet may cause a bug, where the numbers on both sides of the sheet will flip places. 

Revised the Generals section, using the format previously outlined in the to-do list. The Generals section on the Overlord's sheet now allows the Overlord player to record the number of stats the General has, the traits the General has, and there is a miscellaneous notes section for the General. This format should apply to the Leaders that the Empire playbook uses. 

Added new goal for the To-do list: Modify the Spells section of the Fellowship playbook to accommodate the Archmage destiny playbook. One of the custom moves, Enlightened, allows players to use each spell twice. The goal is to add a second checkbox next to each spell to accommodate this.