Update for 5.15.21
Added Maps as a fluid resource for the Fellowship. This is used by the Tinkerer.
Revised the Generals section on the Overlord playbook. This now includes a stat counter similar to the Fellowship's Companions section, as well as a section for writing a General's traits. This format should accommodate the Leaders section of the Empire playbook, when it is developed. 
Updated the To-do list in the Read-me, which contains more info about this update.